nixcon: Tune the sponsorship approval process
Before, there was the implied requirement to wait on organiser objections
for 1 week, with no way to skip if the sponsor is uncontroversial.

This commit refines it to be clear that organisers should issue
objections within 2 days, and only if any objections are issued within
that time, the SC needs to reevaluate whether the sponsor can be
approved.

This allows speeding up the sponsor approval process.