rulesets/nixpkgs: delete accidentally created protected branches
These branches had been created with a protected prefix in the past -
and thus accidentally became protected. This also meant that no pushes
were possible eventually - and thus these branches were quickly
abandoned.

This PR intends to add these temporarily to the ruleset as exceptions -
to then delete the branches immediately. Thus, this PR doesn't need to
actually be merged. Importing the ruleset, deleting the branches, and
then reverting the ruleset to its previous state should be enough.