rulesets/nixpkgs: enable merge queue on secondary development branches
This enables Merge Queues on secondary development branches like staging
and haskell-updates as well.

These branches still need to be pushed to manually for the periodic
merges, both automated by CI and manually by committers. Compared to the
ruleset for merge queues on the primary development branches, this
`always` allows bypassing. Since the number of merges on these branches
is much lower, an occasional bypassing of the queue for the periodic
merges is not a problem, at least resource-wise.

We will still want to eventually implement a PR-based workflow for
periodic merges to make sure these never break the branch.