nixpkgs-committers: hand over to the Nixpkgs core team
Per discussion with the current Nixpkgs committer delegation team
about potential reforms for the commit bit process, they currently
have insufficient time to work through applications and are happy
for the Nixpkgs core team to take over. We’ve decided it’s best
to step in to clear the backlog with the current process, without
blocking on any further changes for now.

We’ll need an org owner to make the following changes:

* Replace @jtojnar and @NickCao with @alyssais, @emilazy, and
  @wolfgangwalther as maintainers of the [`@NixOS/nixpkgs-committers`]
  team.
  
* Replace @Mic92, @jtojnar, and @NickCao with @alyssais,
  @emilazy, and @wolfgangwalther as maintainers of the
  [`@NixOS/retired-nixpkgs-contributors`] team.

* Delete the [`@NixOS/commit-bit-delegation`] team as now redundant
  to @NixOS/nixpkgs-core.

[`@NixOS/nixpkgs-committers`]: <https://github.com/orgs/NixOS/teams/nixpkgs-committers>
[`@NixOS/retired-nixpkgs-contributors`]: <https://github.com/orgs/NixOS/teams/retired-nixpkgs-contributors>
[`@NixOS/commit-bit-delegation`]: <https://github.com/orgs/NixOS/teams/commit-bit-delegation>